Changed sidebar communciation to use load functions
