+++ src/routes/(authenticated)/sidebar.ts
@@ -1,0 +1,5 @@
+export type NavigationElement = {
+	name: string
+	icon: ConstructorOfATypedSvelteComponent
+	route: string
+}
